+++ src/main/java/gift/config/KakaoProperties.java
@@ -1,4 +1,4 @@
-package gift.user.config;
+package gift.config;
 
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.context.properties.ConfigurationProperties;
@@ -6,7 +6,10 @@ import org.springframework.boot.context.properties.ConfigurationProperties;
 @ConfigurationProperties("kakao")
 public record KakaoProperties(
     @Value("${client-id}") String clientId,
-    @Value("${redirect-uri}") String redirectUri
+    @Value("${redirect-uri}") String redirectUri,
+    @Value("token-url") String tokenUrl,
+    @Value("user-info-url") String userInfoUrl,
+    @Value("message-url") String messageUrl
 ) {
 
 }
